small-projects/webgpu/requirements.txt
cffi==1.15.1
numpy==1.23.5
Pillow==9.3.0
pycparser==2.21
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
wgpu==0.8.2
